# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/playground/pricing-agents/{agentId}/playground": {
      "post": {
        "operationId": "PlaygroundController_executePlayground",
        "parameters": [
          {
            "name": "agentId",
            "required": true,
            "in": "path",
            "description": "Pricing agent ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Tenant-ID",
            "required": false,
            "in": "header",
            "description": "Tenant ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "checkpointId",
            "required": false,
            "in": "query",
            "description": "Checkpoint ID (optional, uses latest if not provided)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlaygroundExecutionRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Agent executed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaygroundExecutionResponseDto"
                }
              }
            }
          },
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaygroundExecutionResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - invalid input"
          },
          "404": {
            "description": "Pricing agent or checkpoint not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Execute agent with natural language input for playground testing",
        "tags": [
          "playground"
        ]
      }
    },
    "/playground/pricing-agents/{agentId}/demo-conversation": {
      "post": {
        "operationId": "PlaygroundController_generateDemoConversation",
        "parameters": [
          {
            "name": "agentId",
            "required": true,
            "in": "path",
            "description": "Pricing agent ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Tenant-ID",
            "required": false,
            "in": "header",
            "description": "Tenant ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "checkpointId",
            "required": false,
            "in": "query",
            "description": "Checkpoint ID (optional, uses latest if not provided)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Demo conversation generated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DemoConversationResponseDto"
                }
              }
            }
          },
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DemoConversationResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Pricing agent or checkpoint not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Generate a demo conversation for the pricing agent",
        "tags": [
          "playground"
        ]
      }
    },
...
  },
  "info": {
    "title": "Pricing Agent Builder API",
    "description": "API documentation for the Pricing Agent Builder service",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "ChatbotMessage": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "The message text"
          },
          "role": {
            "type": "string",
            "description": "The role of the message sender",
            "enum": [
              "AI",
              "User"
            ]
          }
        },
        "required": [
          "message",
          "role"
        ]
      },
      "PlaygroundExecutionRequestDto": {
        "type": "object",
        "properties": {
          "input": {
            "type": "string"
          },
          "conversation": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChatbotMessage"
            }
          }
        },
        "required": [
          "input"
        ]
      },
      "BacktraceCalculationStepDTO": {
        "type": "object",
        "properties": {
          "operation": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "subTasks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BacktraceCalculationStepDTO"
            }
          }
        },
        "required": [
          "operation",
          "description"
        ]
      },
      "QuoteErrorDTO": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
      "QuoteResultDTO": {
        "type": "object",
        "properties": {
          "total": {
            "type": "number"
          },
          "pricingCalculationBacktrace": {
            "$ref": "#/components/schemas/BacktraceCalculationStepDTO"
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QuoteErrorDTO"
            }
          }
        }
      },
      "PlaygroundExecutionResponseDto": {
        "type": "object",
        "properties": {
          "structuredOrder": {
            "type": "object",
            "description": "The structured order input converted from natural language"
          },
          "functionResult": {
            "description": "The result from executing the pricing function",
            "allOf": [
              {
                "$ref": "#/components/schemas/QuoteResultDTO"
              }
            ]
          },
          "aiMessage": {
            "type": "string",
            "description": "The AI-generated response message for the playground conversation"
          }
        },
        "required": [
          "structuredOrder",
          "functionResult",
          "aiMessage"
        ]
      },
      "DemoConversationResponseDto": {
        "type": "object",
        "properties": {
          "conversationHistory": {
            "description": "Conversation history with previous messages",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChatbotMessage"
            }
          },
          "nextUserMessage": {
            "type": "string",
            "description": "The next user message for order confirmation"
          }
        },
        "required": [
          "nextUserMessage"
        ]
      },
...
    }
  }
}